pyproject.toml
[project]
name = "dmrserver"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "sqlalchemy>=2.0.42",
    "drf-nested-routers>=0.94.2",
    "gunicorn>=23.0.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.12.12",
]

[tool.ruff]
target-version = "py312" 
line-length = 88

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D",
    "E501",
    "ERA001",
    "COM812",
]
fixable = ["ALL"]

[tool.ruff.lint.per-file-ignores]
"**/tests.py" = [
    "S106",      # hardcoded passwords in tests are acceptable
    "PLC0415",   # local imports inside test functions
    "S101",      # assert used in tests
    "D100",      # docstring rules not critical in tests
    "D103",
    "PLR2004",   # magic numbers in tests are fine
    "ANN",       # missing annotations in tests
    "ARG",       # unused/missing arg annotations in tests
]

"**/tests/*.py" = [
    "S106",
    "PLC0415",
    "S101",
    "D100",
    "D103",
    "PLR2004",
    "ANN",
    "ARG",
]

[tool.ruff.format]
preview = true
